+++ src/lib/Operators/EnvironmentVariable.ts
@@ -17,6 +17,8 @@
 import * as cL from '../Constants';
 import type * as dL from '../Declaration';
 
+const OPT_NAME_DEFAULT = 'default';
+
 export interface IEnvVarOperatorOptions {
 
     /**
@@ -26,35 +28,81 @@ export interface IEnvVarOperatorOptions {
      * @default undefined
      */
     defaultValue?: string;
+
+    /**
+     * A custom function to read environment variables.
+     *
+     * This is useful in scenarios where environment variables are read from an
+     * alternative source other than the default `process.env`.
+     *
+     * @default (name) => process.env[name]
+     */
+    readEnv?: (name: string) => string | null | undefined;
 }
 
-class EnvVarInlineOperatorWithDefault implements dL.IInlineOperator {
+type IReadEnvFn = Required<IEnvVarOperatorOptions>['readEnv'];
+
+const DEFAULT_READ_ENV: IReadEnvFn = (n) => process.env[n];
+
+abstract class AbstractEnvVarInlineOperator implements dL.IInlineOperator {
+
+    protected readonly _readEnv: IReadEnvFn;
+
+    public constructor(readEnv: IReadEnvFn) {
+
+        this._readEnv = readEnv;
+    }
+
+    public abstract process(operand: string, ctx: unknown, opts: dL.IOperationOptions): string;
+
+    public abstract processSync(operand: string, ctx: unknown, opts: dL.IOperationOptions): string;
+
+    protected _read(operand: string, opts: dL.IOperationOptions): string | null {
+
+        const envNames = operand.split(',').map((n) => n.trim());
+
+        for (const envName of envNames) {
+            const value = this._readEnv(envName);
+            if (value !== null && value !== undefined) {
+                return value;
+            }
+        }
+
+        return typeof opts[OPT_NAME_DEFAULT] === 'string' ? opts[OPT_NAME_DEFAULT] : null;
+    }
+}
+
+class EnvVarInlineOperatorWithDefault extends AbstractEnvVarInlineOperator {
 
     private readonly _defaultValue: string;
 
-    public constructor(defaultValue: string) {
+    public constructor(
+        defaultValue: string,
+        readEnv: IReadEnvFn
+    ) {
 
+        super(readEnv);
         this._defaultValue = defaultValue;
     }
 
-    public process(operand: string): string {
+    public process(operand: string, _ctx: unknown, opts: dL.IOperationOptions): string {
 
-        return process.env[operand] ?? this._defaultValue;
+        return this._read(operand, opts) ?? this._defaultValue;
     }
 
-    public processSync(operand: string): string {
+    public processSync(operand: string, _ctx: unknown, opts: dL.IOperationOptions): string {
 
-        return process.env[operand] ?? this._defaultValue;
+        return this._read(operand, opts) ?? this._defaultValue;
     }
 }
 
-class EnvVarInlineOperatorWithoutDefault implements dL.IInlineOperator {
+class EnvVarInlineOperatorWithoutDefault extends AbstractEnvVarInlineOperator {
 
-    public process(operand: string): string {
+    public process(operand: string, _ctx: unknown, opts: dL.IOperationOptions): string {
 
-        const value = process.env[operand];
+        const value = this._read(operand, opts);
 
-        if (value === undefined) {
+        if (value === null) {
 
             throw new Error(`Environment variable '${operand}' is not defined.`);
         }
@@ -62,18 +110,26 @@ class EnvVarInlineOperatorWithoutDefault implements dL.IInlineOperator {
         return value;
     }
 
-    public processSync(operand: string): string {
+    public processSync(operand: string, ctx: unknown, opts: dL.IOperationOptions): string {
 
-        return this.process(operand);
+        return this.process(operand, ctx, opts);
     }
 }
 
 /**
  * This operator reads the value of an environment variable.
  *
+ * @since v1.1.0: If multiple variable names are provided, they will be checked
+ * in order, and the value of the first found variable will be used.
+ * @since v1.1.0: An operator option 'default' is added. When provided, it will
+ * be used as the default value if none of the specified environment variables
+ * are found.
+ *
  * @mode inline
  *
  * @syntax `"$[[env:<variable-name>]]"`
+ * @syntax `"$[[env:<variable-name1>,<variable-name2>]]"`
+ * @syntax `"$[[env:<variable-name1>[...<variable-names>]; default=<default-value>]]"`
  */
 export class EnvironmentVariableOperator implements dL.IOperator {
 
@@ -88,10 +144,15 @@ export class EnvironmentVariableOperator implements dL.IOperator {
     public constructor(opts?: IEnvVarOperatorOptions) {
 
         if (opts?.defaultValue !== undefined) {
-            this.modes[cL.EOperatorMode.INLINE] = new EnvVarInlineOperatorWithDefault(opts.defaultValue);
+            this.modes[cL.EOperatorMode.INLINE] = new EnvVarInlineOperatorWithDefault(
+                opts.defaultValue,
+                opts.readEnv ?? DEFAULT_READ_ENV
+            );
         }
         else {
-            this.modes[cL.EOperatorMode.INLINE] = new EnvVarInlineOperatorWithoutDefault();
+            this.modes[cL.EOperatorMode.INLINE] = new EnvVarInlineOperatorWithoutDefault(
+                opts?.readEnv ?? DEFAULT_READ_ENV
+            );
         }
     }
 }